refactor: Phase 1 - Reorganize models into logical subdirectories
- Created subdirectory structure: buffer/, display/, state/
- Moved model files to appropriate subdirectories
- Updated all imports throughout codebase
- Maintained backward compatibility with re-exports
- All 478 unit tests passing
- All 6 integration tests passing

This is pure reorganization with no functional changes.
