+++ src/repl/models/buffer/buffer_char.rs
