+++ src/repl/models/buffer/mod.rs
@@ -1,0 +1,12 @@
+//! Buffer-related models for text storage and manipulation
+
+pub mod buffer_char;
+pub mod buffer_model;
+pub mod request_model;
+pub mod response_model;
+
+// Re-export commonly used types
+pub use buffer_char::{BufferChar, BufferLine, CharacterBuffer};
+pub use buffer_model::{BufferContent, BufferModel};
+pub use request_model::RequestModel;
+pub use response_model::ResponseModel;
