+++ src/repl/models/buffer/response_model.rs
