+++ src/repl/models/display/display_cache.rs
@@ -3,9 +3,9 @@
 //! Provides display line caching for efficient word wrap rendering and cursor positioning.
 //! Maps logical lines to display lines with position tracking for navigation.
 
-use crate::repl::models::buffer_char::BufferLine;
-use crate::repl::models::display_line::DisplayLine;
-use crate::repl::models::geometry::Position;
+use crate::repl::models::buffer::buffer_char::BufferLine;
+use super::display_line::DisplayLine;
+use crate::repl::models::state::geometry::Position;
 use std::collections::HashMap;
 use std::time::Instant;
