+++ src/repl/models/state/logical_position.rs
