+++ src/repl/view_models/pane_state/cursor_basic.rs
@@ -7,7 +7,7 @@
 //! - Virtual column management for Vim-style navigation
 
 use crate::repl::events::{EditorMode, LogicalPosition, PaneCapabilities, ViewEvent};
-use crate::repl::models::geometry::Position;
+use crate::repl::models::state::geometry::Position;
 
 use super::PaneState;
