+++ src/repl/view_models/pane_state/cursor_line.rs
@@ -7,7 +7,7 @@
 //! - Append mode positioning
 
 use crate::repl::events::{EditorMode, LogicalPosition, PaneCapabilities, ViewEvent};
-use crate::repl::models::geometry::Position;
+use crate::repl::models::state::geometry::Position;
 
 use super::PaneState;
