+++ src/repl/view_models/pane_state/mod.rs
@@ -22,7 +22,7 @@
 //! and follows the Single Responsibility Principle.
 
 use crate::repl::events::{EditorMode, LogicalPosition, Pane, PaneCapabilities};
-use crate::repl::models::geometry::{Dimensions, Position};
+use crate::repl::models::state::geometry::{Dimensions, Position};
 use crate::repl::models::{BufferModel, DisplayCache};
 use crate::repl::view_models::Selection;
 use std::ops::{Index, IndexMut};
