+++ src/repl/view_models/pane_state/scrolling.rs
@@ -7,7 +7,7 @@
 //! - Mode-aware scrolling behavior
 
 use crate::repl::events::{EditorMode, LogicalPosition, PaneCapabilities, ViewEvent};
-use crate::repl::models::geometry::Position;
+use crate::repl::models::state::geometry::Position;
 
 use super::{CursorMoveResult, PaneState, ScrollAdjustResult, ScrollResult};
