+++ src/repl/view_models/pane_state/word_navigation.rs
@@ -7,7 +7,7 @@
 //! - Support for Japanese and multi-byte character word boundaries
 
 use crate::repl::events::{EditorMode, LogicalPosition, PaneCapabilities, ViewEvent};
-use crate::repl::models::geometry::Position;
+use crate::repl::models::state::geometry::Position;
 
 use super::{OptionalPosition, PaneState};
